traffic-signal-optimization/edge/requirements.txt
# Edge Device Requirements - Traffic Signal Optimization
# Python 3.10+

# ============================================
# Computer Vision & Deep Learning
# ============================================
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0

# Deep Learning
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.220  # YOLOv8

# ============================================
# Video Processing
# ============================================
av==10.0.0  # PyAV for RTSP streams
imageio==2.31.5
imageio-ffmpeg==0.4.9

# ============================================
# Communication
# ============================================
paho-mqtt==1.6.1
protobuf==4.24.4
grpcio==1.59.3
grpcio-tools==1.59.3

# ============================================
# Storage & Caching
# ============================================
influxdb-client==1.38.0
redis==5.0.1
hiredis==2.2.3

# ============================================
# Configuration & Environment
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# ============================================
# Utilities
# ============================================
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3  # Retry logic
cachetools==5.3.2

# ============================================
# Async & Concurrency
# ============================================
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.1
aiokafka==0.8.1

# ============================================
# Monitoring & Logging
# ============================================
prometheus-client==0.19.0
loguru==0.7.2
colorlog==6.8.0

# ============================================
# Scientific Computing
# ============================================
scipy==1.11.4
pandas==2.1.3
scikit-learn==1.3.2

# ============================================
# Testing & Development
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# ============================================
# Code Quality
# ============================================
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3